fix: rework services with new common types

+++ services/device-service/src/controller/device-link/delete-device-link.ts
@@ -2,6 +2,7 @@ import Joi from "joi";
 import { ControllerResponse } from "@lindorm-io/koa";
 import { ServerKoaController } from "../../types";
 import { destroyDeviceLinkCallback } from "../../handler";
+import { ClientError } from "@lindorm-io/errors";
 
 interface RequestData {
   id: string;
@@ -19,7 +20,12 @@ export const deleteDeviceLinkController: ServerKoaController<RequestData> = asyn
   const {
     entity: { deviceLink },
     repository: { deviceLinkRepository },
+    token: { bearerToken },
   } = ctx;
+
+  if (deviceLink.identityId !== bearerToken.subject) {
+    throw new ClientError("Invalid bearer token");
+  }
 
   await deviceLinkRepository.destroy(deviceLink, destroyDeviceLinkCallback(ctx));
 };
